requirements.txt
beautifulsoup4
blinker
bs4
certifi
charset-normalizer
click
colorama
dnspython
Flask
idna
itsdangerous
Jinja2
MarkupSafe
pymongo
python-dotenv
requests
soupsieve
urllib3<2.0
Werkzeug
